+++ source/calculators/mcc.hpp
@@ -17,7 +17,7 @@ class MCC : public ClusterMetric  {
     
 public:
     MCC() : ClusterMetric("mcc") {};
-    double getValue( long long tp,  long long tn,  long long fp,  long long fn);
+    double getValue(double tp,  double tn,  double fp,  double fn);
     string getCitation() { return "http://www.mothur.org/wiki/MCC"; }
     
 private:
